split trigger status parsing from SDK for client bundling
