+++ .gitignore
@@ -62,3 +62,4 @@ lib/emails/marketing
 
 # changelog
 changelog
+.env*
